tests/wpt/tests/css/css-properties-values-api/animation/custom-property-animation-inherited-used-by-standard-property.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.css-houdini.org/css-properties-values-api-1">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/utils.js"></script>
<div id="container">
    <div id="target"></div>
</div>
<script>

test(() => {
  CSS.registerProperty({
    name: "--my-length",
    syntax: "<length>",
    inherits: true,
    initialValue: "0px"
  });

  target.style.marginLeft = "var(--my-length)";

  const duration = 1000;
  const animation = container.animate({ "--my-length": "100px" }, duration);
  animation.pause();
  animation.currentTime = duration / 4;

  assert_equals(getComputedStyle(target).marginLeft, "25px");
}, "Animating an inherited CSS variable on a parent is reflected on a standard property using that variable as a value on a child");

</script>
